requirements.txt
Django>=3.0,<4.0
djangorestframework==3.12.4
markdown==3.2.2
pymongo[srv]==3.11.4
mongoengine==0.23.1
django-rest-framework-mongoengine==3.4.1
